added routes fofr delete message and update user. Also updated messages controller with deleteMessage
