+++ controllers/users.js
@@ -96,4 +96,28 @@ const changePassword = async (req, res) => {
   }
 }
 
-module.exports = { signUp, signIn, verify, changePassword }
+const getUsers = async (req, res) => {
+  try {
+      const users = await User.find({});
+      return res.status(200).json(users);
+  } catch (error) {
+      return res.status(500).json({ error: error.message });
+  }
+};
+
+const getUser = async (req, res) => {
+  try {
+      const user = await User.findById(req.params.id);
+      if (user) {
+          return res.status(200).json(user);
+      } else {
+          return res
+              .status(404)
+              .send("Item with specified ID does not exist!");
+      }
+  } catch (error) {
+      return res.status(500).json({ error: error.message });
+  }
+};
+
+module.exports = { signUp, signIn, verify, changePassword, getUsers, getUser }
